separated UI, Map, Tower and added sample Tower image
